[FIX] account_financial_report_webkit: Don't overpass top level account
With current code, the top level accounts (the 0 one) is ignored in
the population of debit, credit and balance data, provoking that the
report shows a line with all 0's, which leads to confussion to users.

I see no reason for zeroing this data, as it can be significant for
certain account charts, and can be easily ignored for those who not.