add an option in partner balance for non-zero only
This options exists in the partner balance from the account module. It
was not there in the webkit module though. The naming is the same as
from the odoo module.

A very old commit message for the webkit report mentions this
functionality but apparently it is not there.

By default all partners are shown as before. This option works for both
PDF and XLS reports.