[8.0] Fix account.move.line field last_rec_date
In previous implementation of the field with the api 7.0, the field was not always computed at the right moment.
This would cause last_rec_date to be null. The drawback is that the amounts on the aged partner balance may be incorrect.
For example, if you have an invoice on May and a payment on June, the amount would not appear on the partner
aged balance of May.