[IMP] Allow to attach from compose wizard
By default we checked for an empty data dict in _get_report_values
however the mail template sends a file type and editor key by default.

We now check if the required company_id is in the dictionary. We also support
supplying force_company in context for setting the company for auto generated
emails.