[FIX] Ensure onchanges applied when calling report directly
Prior to this commit create was called in the report in order to
get the required data. However date_end was not defaulted and relied
on an onchange to get the correct date.

In order to fix we use the new Odoo tests Form functionality to
emulate calling the wizard form just as in UI.