[FIX] account_financial_report: allow navigation on all fields
Pass res_id correctly to lines to allow navigation on the fields

In the open items ledger, the res_id attribute of the invoice
in the report was getting: (id, move_name)
The result was that, when clicking the line, it would redirect
to a new record, instead of the existing.
This passes only the id to the line, solving that issue.

Fix menu item name

TT29371