[IMP] account_sale_stock_report_non_billed: Take partner from picking
With the module account_purchase_stock_report_non_billed, when a
purchase is made and it is received, the supplier is not appearing in
the report of non-billed movements, as it did in previous versions.

So we retrieved the field picking_partner_id to have the same value as
in previous versions.